requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.9
django-crispy-forms==2.1
django-import-export==3.3.7
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
